% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate.R
\name{add_candidate_ids}
\alias{add_candidate_ids}
\title{Add candidate identifiers to precinct returns}
\usage{
add_candidate_ids(.data)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Merges candidate IDs in the \code{[candidates]} dataset, by \code{state}, \code{district},
\code{office}, and a normalizing transformation of \code{candidate}. (This remains in
the data afterward as \code{candidate_normalized}.)
}
