% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{add_county_ids}
\alias{add_county_ids}
\title{Add county identifiers to precinct returns}
\usage{
add_county_ids(.data, must_work = TRUE)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Add county identifiers to precinct returns
}
