% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{add_geo_ids}
\alias{add_geo_ids}
\title{Add state and county identifiers to precinct returns}
\usage{
add_geo_ids(.data, must_work = TRUE)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Add state and county identifiers to precinct returns
}
