% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{add_state_ids}
\alias{add_state_ids}
\title{Add state identifiers to precinct returns}
\usage{
add_state_ids(.data)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Given the state postal code, add the full name of the state and its FIPS and
ICPSR codes.
}
