% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{all_na}
\alias{all_na}
\title{Predicate for all-\code{NA} rows or columns in tidy excel tables}
\usage{
all_na(.data, .by = "row", i = TRUE, j = TRUE)
}
\description{
Predicate for all-\code{NA} rows or columns in tidy excel tables
}
