% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{as_header}
\alias{as_header}
\title{Create identifier columns in spreadsheet data}
\usage{
as_header(.data, idcol, rows = TRUE, cols = TRUE, right = FALSE,
  down = FALSE, .drop = TRUE)
}
\arguments{
\item{.data}{A table of tidy spreadsheet data (i.e., with columns \code{row}, \code{col}, and \code{value}).}

\item{idcol}{String giving the name of the new identifier column.}

\item{rows}{One of: a vector of spreadsheet row indexes; a logical vector with
length equal to the row count \emph{of dataframe \code{.data}}; or a length-one logical
vector (i.e., \code{TRUE} or \code{FALSE}) applied to all spreadsheet rows.}

\item{cols}{One of: a vector of spreadsheet column indexes; a logical vector with
length equal to the column count \emph{of dataframe \code{.data}}; or a length-one logical
vector (i.e., \code{TRUE} or \code{FALSE}) applied to all spreadsheet columns}

\item{right}{If \code{TRUE}, carry non-missing values in cells selected by \code{rows} and
cols\code{rightward along spreadsheet rows. Can be used together with}down`.}

\item{down}{If \code{TRUE}, carry non-missing values in cells selected by \code{rows} and
cols\code{downward along spreadsheet rows. Can be used together with}up`.}

\item{.drop}{If \code{TRUE}, drop spreadsheet cells selected by \code{rows} and \code{cols} after
creating the new identifier column.}
}
\description{
See the Excel vignette for use.
}
