% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{assign_defaults}
\alias{assign_defaults}
\title{Assign default values to missing precinct return variables}
\usage{
assign_defaults(.data, state_postal)
}
\arguments{
\item{.data}{A table of returns.}

\item{state_postal}{State postal abbreviation.}
}
\description{
Create and fill with reasonable defaults the variables \code{writein}, \code{year},
\code{stage}, \code{special}, and \code{mode}.
}
