% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse.R
\name{assign_fixed}
\alias{assign_fixed}
\title{Assign dataverse identifiers by expected values of \code{office}.}
\usage{
assign_fixed(.data)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Assign dataverse identifiers by expected values of \code{office}.
}
