% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse.R
\name{assign_match}
\alias{assign_match}
\title{Assign rows matching an \code{office} pattern to a dataverse}
\usage{
assign_match(.data, pattern, dv, ignore_case = TRUE, ...)
}
\arguments{
\item{.data}{A table of returns.}

\item{pattern}{Pattern to modify behaviour.}

\item{dv}{A value to assign the \code{dataverse} variable for rows in \code{.data} that
match \code{pattern} in column \code{office}.}

\item{ignore_case}{Should case differences be ignored in the match?}

\item{...}{Further arguments to \code{grepl}}
}
\description{
Assign rows matching an \code{office} pattern to a dataverse
}
