% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate.R
\name{candidate_merge_name}
\alias{candidate_merge_name}
\title{Create a merge key from candidate names}
\usage{
candidate_merge_name(.data)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Apply a normalizing transformation of the \code{candidate} column, creating
\code{candidate_normalized}. This is often a lower-cased last name.
}
