% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_candidates.R
\docType{data}
\name{candidates}
\alias{candidates}
\title{A lookup table for candidate identifiers}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 10814 rows and 21 columns.}
\usage{
candidates
}
\description{
A lookup table for candidate identifiers
}
\keyword{datasets}
