% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{carry_forward}
\alias{carry_forward}
\alias{carry_right}
\alias{carry_down}
\title{Carry spreadsheet values forward row-wise or column-wise}
\usage{
carry_forward(.data, rows = TRUE, cols = TRUE, .by = "row",
  value.name = "value")

carry_right(.data, rows = TRUE, cols = TRUE, value.name = "value")

carry_down(.data, rows = TRUE, cols = TRUE, value.name = "value")
}
\arguments{
\item{.data}{A table of tidy spreadsheet data (i.e., with columns \code{row}, \code{col}, and \code{value}).}

\item{rows}{One of: a vector of spreadsheet row indexes; a logical vector with
length equal to the row count \emph{of dataframe \code{.data}}; or a length-one logical
vector (i.e., \code{TRUE} or \code{FALSE}) applied to all spreadsheet rows.}

\item{cols}{One of: a vector of spreadsheet column indexes; a logical vector with
length equal to the column count \emph{of dataframe \code{.data}}; or a length-one logical
vector (i.e., \code{TRUE} or \code{FALSE}) applied to all spreadsheet columns}

\item{.by}{Carry forward by \code{"row"} or by \code{"col"}.}

\item{value.name}{Column in \code{.data} to operate on.}
}
\description{
\code{carry_forward} implements row-wise or column-wise last non-missing
observation carried forward, in a table of tidy Excel data created by
\code{\link[=read_xlreturns]{read_xlreturns()}} or \code{\link[tidyxl:xlsx_cells]{tidyxl::xlsx_cells()}}. Usually called by
\code{\link[=as_header]{as_header()}}, not directly.
}
