% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geo.R
\docType{data}
\name{census_precincts}
\alias{census_precincts}
\title{A table describing Census voting districts}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 177665 rows and 8 columns.}
\source{
\url{https://www.census.gov/geo/maps-data/data/nlt.html}
}
\usage{
census_precincts
}
\description{
A table describing Census voting districts
}
\keyword{datasets}
