% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{col_values}
\alias{col_values}
\alias{su}
\title{Inspect variables' unique values}
\usage{
col_values(.data, n_max = 1000)

su(x)
}
\arguments{
\item{.data}{A table of returns.}

\item{n_max}{A limit on the number of unique values to return. \code{NULL} or \code{0}
to disable.}

\item{x}{A vector.}
}
\description{
Inspect variables' unique values
}
\section{Functions}{
\itemize{
\item \code{su}: A shortcut for sort-unique
}}

