% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{collapse_cells}
\alias{collapse_cells}
\title{Collapse cells}
\usage{
collapse_cells(.data, i, f = paste_omit_na, collapse = " ", .drop = TRUE)
}
\arguments{
\item{.data}{A data.table.}

\item{i}{A vector of row indexes.}

\item{f}{Concatenation (or other) function.}

\item{drop}{Drop all but the first row.}
}
\description{
Concatenate rows column-wise into a single row.
}
