% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{combine_value_cols}
\alias{combine_value_cols}
\title{Combine value columns in tidyxl tables}
\usage{
combine_value_cols(.data)
}
\arguments{
\item{.data}{A table of tidy spreadsheet data (i.e., with columns \code{row}, \code{col}, and \code{value}).}
}
\description{
Creates a new column, \code{value}, taking the values for each observation of
column \code{numeric}, \code{logical}, or \code{character}, whichever is specified in
\code{data_type}.
}
