% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{compare_aggregates}
\alias{compare_aggregates}
\title{Compare vote totals from precinct- and constituency-level returns}
\usage{
compare_aggregates(.data)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Compare vote totals from precinct- and constituency-level returns
}
