% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_votes}
\alias{compare_votes}
\title{Compare vote totals}
\usage{
compare_votes(.data, reference = NULL, by = c("state_postal", "office",
  "candidate"), all.y = FALSE)
}
\arguments{
\item{.data}{Precinct returns dataframe.}

\item{reference}{A dataframe of reference vote totals. See details.}

\item{by}{A vector of shared column names in \code{.data} and \code{reference} to merge
on.}

\item{all.y}{Perform a full join, instead of a left join. This will include
in the result all rows from the reference dataset.}
}
\value{
A copy of \code{.data} with state-candidate vote totals and their
difference from the reference totals.
}
\description{
Compare vote totals with a reference dataset. By default, the reference
totals are the \link{state_totals} dataset from \href{https://uselectionatlas.org}{Dave Leip's Atlas of U.S.Presidential Elections}, for \code{candidate} values
of Donald Trump, Hillary Clinton, Gary Johnson, and "Total" (the sum of votes
for these three and all other candidates).
}
\details{
We sum \code{votes} in each dataframe by the variables specified in argument \code{by},
and then left-join \code{.data} and the reference dataframe on the same variables.
}
\examples{
# Use default reference totals from https://uselectionatlas.org
.data = data.frame(office = 'US President', candidate = c('Gary Johnson',
  'Donald Trump'), state_postal = 'AL', votes = c(44000, 1318200))
compare_votes(.data)

# Use an arbitrary dataframe for reference totals
reference = data.frame(office = 'US President', candidate = c('Gary Johnson',
  'Donald Trump'), state_postal = 'AL', votes = c(43022, 1327200))
compare_votes(.data, reference)
}
