% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{drop_by_predicate}
\alias{drop_by_predicate}
\title{Drop tidyxl rows via predicate functions}
\usage{
drop_by_predicate(d, drop_predicates)
}
\description{
Example uses are dropping spreadsheet rows beyond some (spreadsheet) row
number, or dropping spreadsheet rows in which any cell's values match
a pattern like \code{'total'}.
}
\examples{
library(data.table)
d = data.table(row = 1:4, character = c('a', 'b', 'c', 'total'))
d
drop_predicates = list(
  function(d) { d$row == 3 },
  function(d) { d$character \%=\% 'total' })
# predicate 1 -> drop rows 3
# predicate 2 -> drop 'total' row
drop_by_predicate(d, drop_predicates)
}
