% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse.R
\name{dvna}
\alias{dvna}
\title{Test for \code{office} values not yet assigned to a dataverse}
\usage{
dvna(.data)
}
\arguments{
\item{.data}{A table of returns.}
}
\value{
If successful, returns \code{NULL}; otherwise, prints unassigned values
of \code{office} and stops.
}
\description{
Test for \code{office} values not yet assigned to a dataverse
}
