% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geo.R
\docType{data}
\name{eavs}
\alias{eavs}
\title{The 2016 Election Administration Voting Survey}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 6463 rows and 4 columns.}
\source{
\url{https://www.eac.gov/research-and-data/2016-election-administration-voting-survey}
}
\usage{
eavs
}
\description{
The 2016 Election Administration Voting Survey
}
\seealso{
\link{state_eavs}
}
\keyword{datasets}
