% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{expand_colnames}
\alias{expand_colnames}
\title{Expand headers to include values in initial rows}
\usage{
expand_colnames(.data, i, f = paste_omit_na, collapse = " ", drop = TRUE)
}
\description{
Expand headers to include values in initial rows
}
\examples{
d = data.frame(A = c('B', 'C'), L = c('M', 'N'), stringsAsFactors = FALSE)
d = expand_colnames(d, 1:2)
d
}
