% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party.R
\name{expand_party_abbr}
\alias{expand_party_abbr}
\title{Expand common party abbreviations}
\usage{
expand_party_abbr(party)
}
\arguments{
\item{party}{A vector of party abbreviations.}
}
\description{
TODO: generalize, by moving pattern-substitution pairs to a data file.
}
