% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{extract_headers}
\alias{extract_headers}
\title{Extract nested headers}
\usage{
extract_headers(.data, varname, pattern)
}
\arguments{
\item{.data}{A data.table, typically the result of \code{setDT(read_excel())}.}

\item{varname}{The name to assign the header.}

\item{pattern}{A regular expression matching header values. Case is ignored.}
}
\description{
Extract nested headers
}
\examples{
d = data.table(c(NA, 'precinct', 'A', 'B'),
 c('Assembly', 'John Doe', '1', '2'),
 c(NA, 'Jane Doe', '3', '4'))

extract_headers(d, 'office', 'assembly')
# The matching header is in row 1, column 2. Based on the missingness in
# row 1, column 3, the header applies both to columns 2 and 3.
}
