% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{field_schema}
\alias{field_schema}
\title{Retrieve field schema for a single field}
\usage{
field_schema(name)
}
\arguments{
\item{name}{The name of a field.}
}
\description{
Retrieve field schema for a single field
}
