% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\docType{data}
\name{fields}
\alias{fields}
\title{Field schema in datapackage format}
\format{An object of class \code{list} of length 14.}
\usage{
fields
}
\description{
Field schema in datapackage format
}
\keyword{datasets}
