% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{fill_new_precincts}
\alias{fill_new_precincts}
\title{Fill missing data for new precincts}
\usage{
fill_new_precincts(.data)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Fill missing data for new precincts
}
