% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{fill_writein_candidate}
\alias{fill_writein_candidate}
\title{Assign write-ins the \code{candidate} value of \code{['write-in']}}
\usage{
fill_writein_candidate(.data)
}
\description{
Assign write-ins the \code{candidate} value of \code{['write-in']}
}
