% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{finalize}
\alias{finalize}
\title{Finalize spreadsheet data}
\usage{
finalize(.data)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Drop columns in tidy spreadsheet data, and rename \code{values} as \code{votes}.
}
