% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{\%e\%}
\alias{\%e\%}
\title{Extract pattern matches}
\usage{
string \%e\% pattern
}
\arguments{
\item{string}{A character vector.}

\item{pattern}{A regular expression.}
}
\description{
This is an infix operator for case-insensitive substring extraction via
regex. Matching is non-greedy, i.e., only the first match is returned.
}
\details{
It's implemented as  \code{str_extract(x, regex('pattern', TRUE))} and like
\code{sub(x, 'inverse-pattern', '', ignore.case = TRUE)}.
}
\examples{
'APPLE' \%e\% '\\\\w{3}' == 'APP'
'BANANA' \%e\% 'a' == 'A'
}
