% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{\%-\%}
\alias{\%-\%}
\alias{\%r\%}
\title{Remove pattern matches}
\usage{
string \%-\% pattern

string \%r\% pattern
}
\arguments{
\item{string}{A character vector.}

\item{pattern}{A regular expression.}
}
\description{
Remove a pattern match from a string. Think of of this as regex-powered
string subtraction. Matching is case-insensitive and greedy, i.e., all
matches will be removed.
}
\details{
The equivalent alternatives are \code{str_remove_all(x, regex(pattern, ignore_case = TRUE))} (which this function wraps) or \code{gsub('pattern', '', x, ignore.case = TRUE)}.
}
\section{Functions}{
\itemize{
\item \code{\%r\%}: Alias for removing a pattern match
}}

\examples{
'APPLE' \%-\% 'l\\\\w' == 'APP'
'BANANA' \%-\% 'a|n' == 'B'
}
