% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_heda.R
\docType{data}
\name{heda_codes}
\alias{heda_codes}
\title{A lookup table for Harvard Election Data Archive codes}
\format{An object of class \code{list} of length 4.}
\usage{
heda_codes
}
\description{
This dataset maps election variable codes in the
\href{https://dataverse.harvard.edu/file.xhtml?fileId=2456568&datasetVersionId=42865}{2014 HEDA documentation}
to values used in MEDSL returns.
}
\details{
The naming convention for measure variables in HEDA tables is roughly
\code{eYYYY_OOO*_vv*}, where \code{e} gives election characteristics (e.g., general or
special); \code{YYYY} is a year; \code{OOO*} is a 3- or 4-letter office abbreviation;
and \code{vv*} is a suffix with other characteristics including party and vote
mode.
}
\keyword{datasets}
