% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{jurisdiction_merge_name}
\alias{jurisdiction_merge_name}
\title{Create a merge key for jurisdictions}
\usage{
jurisdiction_merge_name(.data)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Create a merge key for jurisdictions
}
