% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{keep_minimal}
\alias{keep_minimal}
\title{Subset columns in a tidyxl table to those typically useful}
\usage{
keep_minimal(.data, types = FALSE)
}
\description{
Subset columns in a tidyxl table to those typically useful
}
