% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{melt_votes}
\alias{melt_votes}
\title{Melt candidate votes}
\usage{
melt_votes(data, ..., na.rm = TRUE, value.name = "votes",
  variable.name = "candidate", variable.factor = FALSE,
  value.factor = FALSE)
}
\description{
A wrapper for \code{[melt]} in the common case where candidates are in columns and
cells give votes.
}
