% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{merge_census}
\alias{merge_census}
\title{Add Census identifiers to precinct returns}
\usage{
merge_census(.data, state_postal = NULL, must_work = TRUE)
}
\arguments{
\item{.data}{A table of returns.}

\item{state_postal}{State postal abbreviation.}

\item{must_work}{Stop if any jurisdictions aren't matched to FIPS codes.}
}
\description{
Add Census identifiers to precinct returns
}
