% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{normalize_district_numbers}
\alias{normalize_district_numbers}
\title{Drop leading zeroes from district numbers-as-strings}
\usage{
normalize_district_numbers(x)
}
\arguments{
\item{x}{A vector of district identifiers.}
}
\description{
The \code{district} column is expected to be type \code{character}, but where its
values are entirely numeric, we would like these numbers to be formatted
consistently, especially for merging purposes (see \code{[add_candidate_ids]}).
}
