% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party.R
\name{normalize_party}
\alias{normalize_party}
\alias{expand_party_abbr}
\title{Expand common party abbreviations}
\usage{
normalize_party(.data)

expand_party_abbr(party)
}
\arguments{
\item{.data}{A table of returns.}

\item{party}{A vector of party abbreviations.}
}
\description{
Expand common party abbreviations
}
\section{Functions}{
\itemize{
\item \code{expand_party_abbr}: The same operation, for a vector.
}}

