% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate.R
\name{normalize_presidential_candidates}
\alias{normalize_presidential_candidates}
\title{Normalize names of presidential candidates}
\usage{
normalize_presidential_candidates(.data)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Limited to rows where \code{office == 'US President'}.
}
\details{
TODO: generalize, by parameterizing office and moving pattern-substitution
pairs to a data file.
}
