% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{normalize_whitespace}
\alias{normalize_whitespace}
\title{Normalize whitespace in character columns}
\usage{
normalize_whitespace(.data, inner = TRUE)
}
\arguments{
\item{.data}{A table of returns.}

\item{inner}{Whether to normalize inner whitespace.}
}
\description{
Replaces 1+ whitespace characters with one space; trims leading and
trailing whitespace; and replaces zero-length strings with \code{NA}.
}
