% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{not_in_eavs}
\alias{not_in_eavs}
\title{Display \code{jurisdiction} merge keys in precinct returns but not EAVS}
\usage{
not_in_eavs(.data, eavs, must_work = TRUE)
}
\arguments{
\item{.data}{A table of returns.}

\item{eavs}{A table of geo IDs from the \link{eavs} dataset.}

\item{must_work}{Whether to stop if any jurisdictions in \code{eavs} cannot be
matched with those in \code{.data}.}
}
\description{
Display \code{jurisdiction} merge keys in precinct returns but not EAVS
}
