% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{not_in_state}
\alias{not_in_state}
\title{Display \code{jurisdiction} merge keys in EAVS but not precinct returns}
\usage{
not_in_state(.data, eavs, must_work = TRUE)
}
\arguments{
\item{eavs}{A table of EAVS data.}
}
\description{
Display \code{jurisdiction} merge keys in EAVS but not precinct returns
}
