% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse.R
\name{ogrep}
\alias{ogrep}
\title{Test an \code{office} pattern for assigning rows to a dataverse}
\usage{
ogrep(.data, pattern, ignore_case = TRUE, ...)
}
\arguments{
\item{.data}{A table of returns.}

\item{pattern}{Pattern to modify behaviour.}

\item{ignore_case}{Should case differences be ignored in the match?}

\item{...}{Further arguments to \code{grepl}}
}
\description{
Test an \code{office} pattern for assigning rows to a dataverse
}
