% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{paste_omit_na}
\alias{paste_omit_na}
\title{Concatenate strings omitting NAs}
\usage{
paste_omit_na(..., sep = " ", collapse = NULL)
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}

\item{sep}{a character string to separate the terms.  Not
    \code{\link{NA_character_}}.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link{NA_character_}}.}
}
\description{
Concatenate strings omitting NAs
}
