% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\docType{data}
\name{precinct_column_types}
\alias{precinct_column_types}
\title{Expected column types in MEDSL precinct returns}
\format{An object of class \code{character} of length 37.}
\usage{
precinct_column_types
}
\description{
Generated from \code{[precinct_validity]}.
}
\seealso{
\code{[precinct_validity]}
}
\keyword{datasets}
