% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\docType{data}
\name{precinct_columns}
\alias{precinct_columns}
\title{Expected column names in MEDSL precinct returns}
\format{An object of class \code{character} of length 38.}
\usage{
precinct_columns
}
\description{
Generated from \code{[precinct_validity]}.
}
\keyword{datasets}
