% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\docType{data}
\name{precinct_validity}
\alias{precinct_validity}
\title{Expected Properties of Columns in MEDSL Precinct Returns}
\format{An object of class \code{list} of length 38.}
\usage{
precinct_validity
}
\description{
Defines a schema for column validity. The name of each list element is a
column that may occur in MEDSL precinct returns. Sub-elements describe
constraints to be imposed on these columns. (This validation is implemented
in \code{[write_precincts}]).
}
\details{
For example, the \code{dataverse} element of \code{precinct_validity} defines the
expected \code{type} of the column named \code{dataverse} in precinct returns:
\code{character}. It should not have \code{NA} values, so \code{no_na = TRUE}. In fact, it
should only take one of the enumerated \code{values}: \code{"president"}, \code{"senate"},
etc.
}
\seealso{
\code{[write_precincts]}
}
\keyword{datasets}
