% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_double_header}
\alias{read_double_header}
\title{Map first-row headers to second-row headers}
\usage{
read_double_header(path, sep = ",", h1_name = "candidate",
  h2_name = "party", ...)
}
\arguments{
\item{path}{Path to a CSV.}
}
\description{
Use in the case where the first row gives candidate names and the second row
gives their party
}
