% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heda.R
\name{read_heda}
\alias{read_heda}
\title{Read and reshape Harvard Election Data Archive returns}
\usage{
read_heda(path)
}
\arguments{
\item{path}{Path to a tabular file of HEDA election returns.}
}
\value{
A dataframe of MEDSL-formatted returns.
}
\description{
Read and melt a file of HEDA returns, then apply known mappings from HEDA
codes to MEDSL variable values. These mappings are from the \href{https://dataverse.harvard.edu/file.xhtml?fileId=2456568&datasetVersionId=42865}{2014documentation}
and not exhaustive. Warnings accompany unknown HEDA codes.
}
