% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{read_spreadsheet}
\alias{read_spreadsheet}
\title{Read an Excel file with multiple header rows}
\usage{
read_spreadsheet(path, top = list(candidate = 1), inner = list(),
  drop = list(), sheet, debug = TRUE)
}
\arguments{
\item{path}{Path to the xlsx file.}

\item{sheet}{Sheet to read. Either a character vector (the name of the sheet)
or an integer (the index of the sheet). If missing, defaults to the first
sheet.}
}
\description{
Deprecated; see new Excel vignette. Still used in ID.
}
\examples{
drop = list(
   function(d) { d$row >= 1130 },
   function(d) { d$character \%=\% 'total' })

# Headers to extract via row number
top = list(office = 1:3, party = 4, candidate = 5)

# Headers to extract via predicate
inner = list(
  jurisdiction = function(d) { d$col == 1 & d$row > 5 & d$local_format_id == 83 },
  precinct = function(d) { d$col == 1 & d$row > 5 & d$local_format_id != 83 })

d = read_spreadsheet(path = '../data-ext/example-returns/16gen_stwd_pct.xlsx',"Federal",
  top, inner, drop)
head(d)

max(d$row)
unique(d$office)
unique(d$party)
unique(d$candidate)

head(unique(d$jurisdiction))
any(is.na(d$jurisdiction))

any(is.na(d$precinct))
head(unique(d$precinct))
}
