% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{read_xlreturns}
\alias{read_xlreturns}
\title{Read Excel spreadsheets}
\usage{
read_xlreturns(path = path, ..., keep_all = FALSE)
}
\description{
Read Excel spreadsheets
}
\examples{
`\%>\%` = magrittr::`\%>\%`
d = read_xlreturns('../data-ext/example-returns/16gen_stwd_pct.xlsx')
d = d \%>\%
  as_idcol('office', i = 1)
}
