% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{rename_legacy_vars}
\alias{rename_legacy_vars}
\title{Rename legacy variable names}
\usage{
rename_legacy_vars(.data)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Rename legacy variable names
}
