% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{row_contains}
\alias{row_contains}
\title{Predicate for rows in tidy excel table containing a pattern}
\usage{
row_contains(.data, pattern, i = TRUE, j = TRUE)
}
\description{
Predicate for rows in tidy excel table containing a pattern
}
