% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{search_all}
\alias{search_all}
\title{Search across all character columns}
\usage{
search_all(.data, pattern, ignore_case = TRUE)
}
\arguments{
\item{.data}{A table of returns.}

\item{pattern}{Pattern to modify behaviour.}

\item{ignore_case}{Should case differences be ignored in the match?}
}
\description{
Search across all character columns
}
