% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{split_cells}
\alias{split_cells}
\title{Split multiple tables in spreadsheets into separate dataframes}
\usage{
split_cells(.data, pattern, i = TRUE, j = TRUE, starts = TRUE)
}
\description{
Split multiple tables in spreadsheets into separate dataframes
}
