% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{split_sheet}
\alias{split_sheet}
\title{Split tables by pattern}
\usage{
split_sheet(path, pattern_column, pattern, adjust_start = 0, adjust_max = 0,
  sheet = NULL, idcol = NULL, use_names = TRUE)
}
\description{
Search for headers that delimit multiple tables in a spreadsheet, and read
the tables separately.
}
