% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_examples.R
\name{spreadsheet_example}
\alias{spreadsheet_example}
\title{Retrieve paths to packaged Excel spreadsheets}
\usage{
spreadsheet_example(name)
}
\arguments{
\item{name}{Name of an Excel file example. See details.}
}
\value{
Path to the example.
}
\description{
The package includes spreadsheets for use in examples and vignettes.
Installation places them in the package directory. This function retrieves
the path to a spreadsheet, given one of the names below.
}
\details{
Spreadsheet names:
\itemize{
\item \code{common-toc}
\item \code{louisiana}
\item \code{merrimack}
\item \code{merrimack-excerpt}
\item \code{nebraska}
\item \code{new-hampshire}
\item \code{new-york-cortland}
\item \code{ohio}
\item \code{oregon}
\item \code{south-dakota}
\item \code{utah-salt-lake}
\item \code{vermont}
\item \code{wisconsin}
}
}
\examples{
# Paths will vary system-to-system
spreadsheet_example('merrimack-excerpt')
}
