% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{standardize_county_fips}
\alias{standardize_county_fips}
\title{Standardize county FIPS codes}
\usage{
standardize_county_fips(.data)
}
\arguments{
\item{.data}{A table of returns.}
}
\description{
Standardize county FIPS codes
}
