% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{state_eavs}
\alias{state_eavs}
\title{Load EAVS precinct data for a given state}
\usage{
state_eavs(state_postal)
}
\arguments{
\item{state_postal}{State postal abbreviation, e.g. \code{"NY"}.}
}
\value{
The \code{[eavs]} dataset subset to \code{state_postal}.
}
\description{
Load EAVS precinct data for a given state
}
