% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\docType{data}
\name{state_totals}
\alias{state_totals}
\title{Presidential vote totals in 2016, by state}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 204 rows and 4 columns.}
\usage{
state_totals
}
\description{
A table of vote totals from \href{https://uselectionatlas.org/RESULTS}{Dave Leip's US Election Atlas}.
}
\keyword{datasets}
