% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\alias{validate_field}
\alias{select_noninteger}
\alias{select_nonstring}
\alias{select_not_enumerated}
\alias{select_missing}
\title{Validate election returns against schema}
\usage{
validate(.data)

validate_field(.data, name)

select_noninteger(.data, name)

select_nonstring(.data, name)

select_not_enumerated(.data, name)

select_missing(.data, name)
}
\arguments{
\item{.data}{A table of election returns.}

\item{name}{The name of a column in \code{.data}.}
}
\description{
Validate election returns against schema
}
\section{Functions}{
\itemize{
\item \code{validate_field}: Validate a single field

\item \code{select_noninteger}: Select non-integer values

\item \code{select_nonstring}: Select non-string values

\item \code{select_not_enumerated}: Select non-enumerated values

\item \code{select_missing}: Select missing values
}}

