% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_precincts}
\alias{write_precincts}
\title{Save returns to disk after validation}
\usage{
write_precincts(.data, state_postal, skip = NULL)
}
\arguments{
\item{.data}{A table of returns.}

\item{state_postal}{State postal abbreviation.}

\item{skip}{Validations in \code{precinct_validity} to skip, as a list whose
names are variables and whose values are properties.}
}
\description{
Save returns to disk after validation
}
\examples{
# 2016 MI precinct returns include statistical adjustments with negative
# `votes`, which requires an exception to the requirement that `votes` values
# are integers >= 0. We might pass:
skip = list(votes = 'is_valid')
# Because the `is_valid` callable in `precinct_validity$votes` enforces the
# >=0 requirement:
data(precinct_validity)
precinct_validity$votes
}
