% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{add_jurisdiction_fips}
\alias{add_jurisdiction_fips}
\title{Merge FIPS codes from EAVS}
\usage{
add_jurisdiction_fips(.data, state_postal, add_key = TRUE, must_work = TRUE)
}
\arguments{
\item{must_work}{Stop if any jurisdictions aren't matched to FIPS codes.}
}
\description{
Merge FIPS codes from EAVS
}
