% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{all_na}
\alias{all_na}
\title{Identify all-\code{NA} rows or columns in spreadsheet}
\usage{
all_na(.data, .by = "row", rows = TRUE, cols = TRUE)
}
\arguments{
\item{.data}{A table of tidy spreadsheet data (i.e., with columns \code{row}, \code{col}, and \code{value}).}

\item{.by}{Either \code{"row"} to find all-NA spreadsheet rows or \code{"col"} to find all-NA
columns.}

\item{rows}{One of: a vector of spreadsheet row indexes; a logical vector with
length equal to the row count \emph{of dataframe \code{.data}}; or a length-one logical
vector (i.e., \code{TRUE} or \code{FALSE}) applied to all spreadsheet rows.}

\item{cols}{One of: a vector of spreadsheet column indexes; a logical vector with
length equal to the column count \emph{of dataframe \code{.data}}; or a length-one logical
vector (i.e., \code{TRUE} or \code{FALSE}) applied to all spreadsheet columns}
}
\description{
Identify all-\code{NA} rows or columns in spreadsheet
}
