% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{as_idcol}
\alias{as_idcol}
\title{Create identifier columns in tidy Excel data}
\usage{
as_idcol(.data, idcol, i = TRUE, j = TRUE, right = FALSE, down = FALSE,
  .drop = TRUE)
}
\description{
See the Excel vignette for use.
}
