% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{assign_defaults}
\alias{assign_defaults}
\title{Assign default values to missing precinct return variables}
\usage{
assign_defaults(.data, state_postal)
}
\description{
Assign default values to missing precinct return variables
}
