% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{carry_down}
\alias{carry_down}
\title{Carry values right in tidy Excel data}
\usage{
carry_down(.data, i = TRUE, j = TRUE, idcol = NULL)
}
\description{
Usually called by \link{as_idcol}, not directly. This is row-wise last
non-missing observation carried forward.
}
