% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\docType{data}
\name{fields}
\alias{fields}
\title{Field schema in datapackage format}
\format{An object of class \code{list} of length 14.}
\usage{
fields
}
\description{
Defines a schema for column validity. The name of each list element is a
column that may occur in MEDSL precinct returns. Sub-elements describe
characteristics of these columns, including constraints imposed.
}
\seealso{
\code{[validate]}
}
\keyword{datasets}
