% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{\%=\%}
\alias{\%=\%}
\title{Detect pattern matches}
\usage{
string \%=\% pattern
}
\arguments{
\item{string}{A character vector.}

\item{pattern}{A regular expression.}
}
\value{
A logical vector.
}
\description{
\code{\%=\%} is made for subsetting dataframes. Matching is case-insensitive
and vectorized over \code{string}.
}
\details{
This is almost identical to \code{\%like\%} in \code{data.table} but it
ignores case by default. We could use \code{grepl} or \code{str_detect}, but
we do a lot of subsetting, and \code{\%=\%} is as many as 30 characters
shorter when ignoring case, while maintaining readability.
}
\examples{
c('apple', 'banana', 'pear') \%=\% 'pp'

data(virginia_precincts)
virginia_precincts[OfficeTitle \%=\% 'house', .(OfficeTitle)]
}
