% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nlapply}
\alias{nlapply}
\title{\code{lapply} with named outputs}
\usage{
nlapply(X, FUN, ..., name_transform = NULL)
}
\arguments{
\item{name_transform}{A function to transform \code{X} before using it as the
output's names.}
}
\description{
This is equivalent to \code{sapply(x, f, simplify = FALSE, USE.NAMES = TRUE)}.
}
