% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{not_in_eavs}
\alias{not_in_eavs}
\title{Display \code{jurisdiction} merge keys in precinct returns but not EAVS}
\usage{
not_in_eavs(.data, eavs, must_work = TRUE)
}
\arguments{
\item{eavs}{A table of EAVS data.}
}
\description{
Display \code{jurisdiction} merge keys in precinct returns but not EAVS
}
