% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{not_in_state}
\alias{not_in_state}
\title{Display \code{jurisdiction} merge keys in EAVS but not precinct returns}
\usage{
not_in_state(.data, eavs, must_work = TRUE)
}
\arguments{
\item{.data}{A table of returns.}

\item{eavs}{A table of geo IDs from the \link{eavs} dataset.}

\item{must_work}{Whether to stop if any jurisdictions in \code{.data} cannot be
matched with those in \code{eavs}.}
}
\description{
Display \code{jurisdiction} merge keys in EAVS but not precinct returns
}
