% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_dir}
\alias{read_dir}
\title{Read a directory of returns data}
\usage{
read_dir(path = "../raw", f = data.table::fread, idcol = "path", ...)
}
\arguments{
\item{path}{Path to a directory of returns.}

\item{f}{A function for reading files from disk.}

\item{idcol}{A name for the column in the result giving file paths.}

\item{...}{Further arguments to \code{\link[base]{list.files}}.}
}
\description{
A wrapper for the pattern:
}
\details{
\itemize{
\item Locate paths with \code{list.files}, using \code{full.names = TRUE} and perhaps a \code{pattern}
\item Apply a read function like \href{data.table}{fread} or
\href{xlreader}{read_excel} to each path
\item Combine the data with e.g. \code{\link{rbindlist}}, using the paths as an identifier
column
}
}
