% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_legacy_csv}
\alias{read_legacy_csv}
\title{Read CSV-formatted returns from the disk}
\usage{
read_legacy_csv(path, strict = TRUE)
}
\arguments{
\item{path}{Path to a CSV.}

\item{strict}{Promote warnings to errors.}
}
\description{
Parser warnings are promoted to errors and legacy variables are renamed.
}
