% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_xlreturns}
\alias{read_xlreturns}
\title{Read spreadsheet data from Excel files}
\usage{
read_xlreturns(path, ..., keep_all = FALSE)
}
\arguments{
\item{path}{Path to the xlsx file.}

\item{...}{Further arguments to \code{\link[tidyxl:xlsx_cells]{tidyxl::xlsx_cells}}}
}
\description{
Read spreadsheet data from Excel files
}
\examples{
d = read_xlreturns('../data-ext/example-returns/16gen_stwd_pct.xlsx')
as_idcol(d, 'office', i = 1)
}
