% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_xlreturns}
\alias{read_xlreturns}
\title{Read spreadsheet data from Excel files}
\usage{
read_xlreturns(path, ..., keep_all = FALSE)
}
\arguments{
\item{path}{Path to the xlsx file.}

\item{...}{Further arguments to \code{\link[tidyxl:xlsx_cells]{tidyxl::xlsx_cells()}}}

\item{keep_all}{Keep all variables returned by \code{\link[tidyxl:xlsx_cells]{tidyxl::xlsx_cells()}},
instead of only those typically needed. (See \code{\link[=keep_minimal]{keep_minimal()}}.)}
}
\description{
Read spreadsheet data from Excel files
}
\examples{
path = spreadsheet_example('wisconsin')
d = read_xlreturns(path)
}
