% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{row_contains}
\alias{row_contains}
\alias{col_contains}
\title{Identify spreadsheet rows that contain a pattern}
\usage{
row_contains(.data, pattern, i = TRUE, j = TRUE)

col_contains(.data, pattern, i = TRUE, j = TRUE)
}
\description{
Identify spreadsheet rows that contain a pattern
}
\section{Functions}{
\itemize{
\item \code{col_contains}: Identify spreadsheet columns that contain a pattern
}}

