% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{split_cells}
\alias{split_cells}
\title{Split multiple tables in spreadsheets into separate dataframes}
\usage{
split_cells(.data, pattern, rows = TRUE, cols = TRUE, starts = TRUE)
}
\arguments{
\item{.data}{A table of tidy spreadsheet data (i.e., with columns \code{row}, \code{col}, and \code{value}).}

\item{pattern}{Pattern to modify behaviour.}

\item{rows}{One of: a vector of spreadsheet row indexes; a logical vector with
length equal to the row count \emph{of dataframe \code{.data}}; or a length-one logical
vector (i.e., \code{TRUE} or \code{FALSE}) applied to all spreadsheet rows.}

\item{cols}{One of: a vector of spreadsheet column indexes; a logical vector with
length equal to the column count \emph{of dataframe \code{.data}}; or a length-one logical
vector (i.e., \code{TRUE} or \code{FALSE}) applied to all spreadsheet columns}

\item{starts}{Whether the pattern indicates the start of a table or the end of the
table.}
}
\description{
Split multiple tables in spreadsheets into separate dataframes
}
