% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_precincts}
\alias{write_precincts}
\title{Save returns to disk}
\usage{
write_precincts(.data, state_postal)
}
\arguments{
\item{.data}{A table of returns.}

\item{state_postal}{State postal abbreviation.}
}
\value{
Processed \code{.data}.
}
\description{
Write data to disk as \code{.csv} and \code{.Rds}, using the naming convention
\code{2016-{state_postal}-precinct}. Transformations are applied before saving;
see details.
}
\details{
Before saving:
\itemize{
\item Rename or drop legacy variables, which appear in some old intermediate
CSVs, e.g., \code{total votes} or \code{candidate.votes}
\item Keep only the columns defined in the \code{[fields]} schema
\item Assign default values for these columns (e.g., \code{FALSE} for \code{special})
\item Normalize whitespace in character columns (trim exterior whitespace;
replace 1+ interior whitespace with single whitespace)
\item Reorder the columns
\item Convert variables to expected types (e.g., numeric -> integer), which
should be non-destructive if the \code{validate(.data)} was successful.
}
}
